Reimplented GameObject velocity, updated and added comments, cleanup
Re-added the velocity property and associated methods to the GameObject classs. Updated relevant areas to use velocity and methods to add to it. Objects are moved when move() is called on then inside update(), consuming their accumulated velocity.
Adjusted platform layout slightly by changing the position of the final platform.
Removed inputs for moving vertically up and down as these were are not used in the final game.
Removed BUFFER_OFFSET macro because it was unused in the code.
Added new comments throughout the codebase and updated existing.